Move service inside AddonTranslator
