+++ nms/src/main/kotlin/io/github/pylonmc/pylon/core/nms/NmsAccessorImpl.kt
@@ -12,14 +12,14 @@ object NmsAccessorImpl : NmsAccessor {
 
     private val players = ConcurrentHashMap<UUID, PlayerPacketHandler>()
 
-    override fun register(player: Player, handler: PlayerTranslationHandler) {
+    override fun registerTranslationHandler(player: Player, handler: PlayerTranslationHandler) {
         if (players.containsKey(player.uniqueId)) return
         val handler = PlayerPacketHandler((player as CraftPlayer).handle, handler)
         players[player.uniqueId] = handler
         handler.register()
     }
 
-    override fun unregister(player: Player) {
+    override fun unregisterTranslationHandler(player: Player) {
         val handler = players.remove(player.uniqueId) ?: return
         handler.unregister()
     }
