+++ pylon-core/src/main/kotlin/io/github/pylonmc/pylon/core/addon/PylonAddonListener.kt
@@ -2,7 +2,7 @@ package io.github.pylonmc.pylon.core.addon
 
 import io.github.pylonmc.pylon.core.block.BlockStorage
 import io.github.pylonmc.pylon.core.entity.EntityStorage
-import io.github.pylonmc.pylon.core.i18n.PylonLanguageService
+import io.github.pylonmc.pylon.core.i18n.AddonTranslator
 import io.github.pylonmc.pylon.core.registry.PylonRegistry
 import org.bukkit.event.EventHandler
 import org.bukkit.event.Listener
@@ -22,7 +22,7 @@ internal object PylonAddonListener : Listener {
             PylonRegistry.RECIPE_TYPES.unregisterAllFromAddon(plugin)
             PylonRegistry.MOB_DROPS.unregisterAllFromAddon(plugin)
             PylonRegistry.ADDONS.unregister(plugin)
-            PylonLanguageService.unregister(plugin)
+            AddonTranslator.unregister(plugin)
         }
     }
 }
